chore(vscode/api): update app bootstrap, package.json, and build output
