--- packages/vscode-extension/src/services/scan.ts
@@ -1,208 +1,0 @@
-import * as vscode from 'vscode';
-import * as fs from 'fs';
-import * as path from 'path';
-import { todosApi } from '../tree/todosApi';
-import {
-  PredefinedTodoTypeEnum,
-  PredefinedTodoTypeEnumType,
-  ProcessedTodos,
-  RawTodo,
-} from '@dev-dashboard/shared';
-
-export const scanTodos = async () => {
-  const startTime = Date.now();
-  console.log('Starting TODO scan...');
-
-  if (!vscode.workspace.workspaceFolders) {
-    vscode.window.showWarningMessage('No workspace folder open');
-    return;
-  }
-
-  const workspaceFolder = vscode.workspace.workspaceFolders[0];
-  console.log('Scanning workspace:', workspaceFolder.uri.fsPath);
-
-  const files = await getSourceFiles(workspaceFolder.uri.fsPath);
-  console.log(`Found ${files.length} files to scan`);
-
-  const todoPromises = files.map(file => scanFileForTodos(file));
-  const todoArrays = await Promise.all(todoPromises);
-  const rawUndefinedTodos: RawTodo[] = todoArrays.flat();
-  const processedTodos: ProcessedTodos[] = processTodos(rawUndefinedTodos);
-
-  console.log(`Found ${processedTodos.length} TODOs:`);
-  processedTodos.forEach(todo => {
-    const tag = todo.type === 'OTHER' ? todo.customTag : todo.type;
-    console.log(
-      `[${tag}] ${todo.content} (${todo.filePath}:${todo.lineNumber})`
-    );
-  });
-
-  try {
-    console.log('Syncing TODOs with API...');
-    await todosApi.syncTodos(processedTodos);
-    console.log('TODOs synced successfully');
-  } catch (error) {
-    console.log(error);
-    console.error('Failed to sync TODOs:', error);
-    if (error instanceof Error) {
-      vscode.window.showErrorMessage(`Failed to sync TODOs: ${error.message}`);
-    }
-  }
-
-  const endTime = Date.now();
-  console.log(`Scan completed in ${endTime - startTime}ms`);
-  vscode.window.showInformationMessage(
-    `Found ${processedTodos.length} TODOs in ${endTime - startTime}ms`
-  );
-};
-
-const getSourceFiles = async (rootPath: string): Promise<string[]> => {
-  const files: string[] = [];
-
-  const skipDirs = [
-    'node_modules',
-    '.git',
-    'dist',
-    'build',
-    'coverage',
-    '.next',
-    '.nuxt',
-    'target',
-    'bin',
-    'obj',
-    '__pycache__',
-    '.pytest_cache',
-    '.venv',
-    'venv',
-  ];
-
-  const skipFiles = [
-    /\.min\.(js|css)$/,
-    /\.bundle\.(js|css)$/,
-    /package-lock\.json$/,
-    /yarn\.lock$/,
-  ];
-
-  const shouldSkipFile = (filename: string): boolean => {
-    return skipFiles.some(pattern => pattern.test(filename));
-  };
-
-  const traverse = async (dir: string): Promise<void> => {
-    try {
-      const items = await fs.promises.readdir(dir);
-
-      const itemPromises = items.map(async item => {
-        const fullPath = path.join(dir, item);
-
-        try {
-          const stat = await fs.promises.stat(fullPath);
-
-          if (stat.isDirectory()) {
-            if (!skipDirs.includes(item)) {
-              await traverse(fullPath);
-            }
-          } else {
-            if (!shouldSkipFile(item)) {
-              files.push(fullPath);
-            }
-          }
-        } catch (error) {
-          console.log(`Error processing ${fullPath}: ${error}`);
-        }
-      });
-
-      await Promise.all(itemPromises);
-    } catch (error) {
-      console.log(`Error reading directory ${dir}: ${error}`);
-    }
-  };
-
-  await traverse(rootPath);
-  return files;
-};
-
-const processTodos = (todos: RawTodo[]): ProcessedTodos[] => {
-  return todos.map(todo => {
-    const isPredefined = PredefinedTodoTypeEnum.safeParse(todo.type).success;
-
-    if (isPredefined) {
-      return {
-        ...todo,
-        type: todo.type as PredefinedTodoTypeEnumType,
-        customTag: undefined,
-      };
-    } else {
-      return {
-        ...todo,
-        type: 'OTHER',
-        customTag: todo.type,
-      };
-    }
-  });
-};
-
-const scanFileForTodos = async (filePath: string): Promise<RawTodo[]> => {
-  const todoLinePatterns = [
-    /(?<!:)\s*\/\/\s*@?([A-Za-z][A-Za-z0-9_-]{0,31})\s*(?:[:\\-]\s*|\s+)(.+)$/i,
-    /^\s*(?:\/\*+|\*)\s*@?([A-Za-z][A-Za-z0-9_-]{0,31})\s*(?:[:\\-]\s*|\s+)(.+)$/i,
-  ];
-
-  const todos: RawTodo[] = [];
-
-  try {
-    const content = await fs.promises.readFile(filePath, 'utf8');
-    const lines = content.split('\n');
-
-    lines.forEach((line, index) => {
-      for (const pattern of todoLinePatterns) {
-        const match = line.match(pattern);
-        if (match) {
-          const rawContent = match[2].replace(/\s*\*\/\s*$/, '');
-          todos.push({
-            type: match[1].toUpperCase(),
-            content: rawContent.trim(),
-            filePath,
-            lineNumber: index + 1,
-          });
-          break;
-        }
-      }
-    });
-  } catch (error) {
-    console.error(`Error reading file ${filePath}: ${error}`);
-    throw error;
-  }
-
-  return todos;
-};
-
-// const scanFileForTodos = async (
-//   filePath: string
-// ): Promise<RawUndefinedTodoBaseSchema[]> => {
-//   const todos: RawUndefinedTodoBaseSchema[] = [];
-
-//   const todoPattern = /\/\/\s*(TODO|FIXME|HACK|NOTE|XXX|BUG)[\s:]*(.*)/gi;
-
-//   try {
-//     const content = await fs.promises.readFile(filePath, 'utf8');
-//     const lines = content.split('\n');
-
-//     lines.forEach((line, index) => {
-//       let match;
-//       todoPattern.lastIndex = 0;
-//       while ((match = todoPattern.exec(line)) !== null) {
-//         todos.push({
-//           type: match[1].toUpperCase(),
-//           content: match[2].trim(),
-//           filePath: filePath,
-//           lineNumber: index + 1,
-//         });
-//       }
-//     });
-//   } catch (error) {
-//     console.error(`Error reading file ${filePath}: ${error}`);
-//     throw error;
-//   }
-
-//   return todos;
-// };
