+++ packages/vscode-extension/src/services/scan/index.ts
@@ -1,0 +1,3 @@
+export * from './get-source-files';
+export * from './scan-file';
+export * from './process-todos';
